requirements.txt
einops==0.6.1
ffmpeg_python==0.2.0
kornia==0.6.11
lpips==0.1.4
numpy==1.21.6
opencv_contrib_python==4.7.0.72
opencv_python==4.7.0.68
PyYAML==6.0
Requests==2.30.0
scipy==1.7.3
setuptools==58.0.0
tabulate==0.9.0
timm==0.6.12
torch==1.10.2+cu113
torchvision==0.11.3+cu113
tqdm==4.64.1
